backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "greenlet>=3.2.3",
    "httpx-oauth>=0.16.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.0",
    "sqlmodel>=0.0.24",
    "stripe>=12.3.0",
    "uvicorn>=0.35.0",
]
